Pass down user object to admin dashboard
