+++ itoapp/src/admin/components/Tasks.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import db from '../../config/firebaseConfig';
-import firebase from 'firebase';
 
 import Task from './Task';
 import AddTaskLink from './AddTaskLink';
@@ -13,64 +12,27 @@ class Tasks extends React.Component {
         this.state = {
             loading: true,
             tasks: [],
-            admin: {},
+            admin: this.props.admin,
         }
     }
-
-    getUser() {
-
-        firebase.auth().onAuthStateChanged((user) => {
-            if (user) { 
-                this.setState({
-                    admin: user,
-                })
-            }
-        })
-
-        
-        
-
-        
-        // firebase.auth().onAuthStateChanged((user) => {
-        //   if (user) {
-    
-        //     const activeuser = db.collection("users").where("authid", "==", admin.uid);
-        //     activeuser.get().then(snapshot => {
-        //       snapshot.forEach(doc => {
-        //         this.setState({
-        //           admin: doc.data()
-        //         })
-                
-        //       })
-        //     })
-        //   } else {
-        //     // User is not signed in. Push to login screen
-        //     console.log("User signed out")
-        //     this.props.history('/login');
-        //   }
-        // });
-      }
 
     getTasks() {
         this.setState({
             loading: false
         })
-        // const tasks = db.collection("tasks").orderBy("createdon");
-        const tasks = db.collection("tasks").where("authid", "==", this.state.admin).orderBy("createdon");
-        console.log(this.state.admin.data())
-        tasks
-            .onSnapshot(snapshot => {
-                const task = snapshot.docs.map(doc => doc.data());
-                // console.log(task);
-                this.setState({ tasks: task });
-            });
+        const tasks = db.collection("tasks").orderBy("createdon");
+
+        // const tasks = db.collection("tasks").where("authid", "==", this.state.admin).orderBy("createdon");
+        // console.log(this.state.admin.data())
+        tasks.onSnapshot(snapshot => {
+            let task = snapshot.docs.map(doc => doc.data());
+            this.setState({ tasks: task });
+        });
 
     }
 
     componentDidMount() {
-        this.getUser();
         this.getTasks();
-        console.log(this.state.admin.uid);
     }
 
     render() {
